Restructure builtin classes
